src: cast strlen() to uint16_t

+++ src/classic/opp_server.c
@@ -206,7 +206,7 @@ void opp_server_create_sdp_record(uint8_t *service, uint32_t service_record_hand
 
     // 0x0100 "Service Name"
     de_add_number(service, DE_UINT, DE_SIZE_16, 0x0100);
-    de_add_data(service, DE_STRING, strlen(name), (uint8_t *) name);
+    de_add_data(service, DE_STRING, (uint16_t) strlen(name), (uint8_t *) name);
 
 #ifdef ENABLE_GOEP_L2CAP
     // 0x0200 "GOEP L2CAP PSM"
